requirements.txt
# Trading Bot Dependencies
binance-connector==3.4.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.3
python-dotenv==1.0.0